requirements.txt
Django==1.8.4
Markdown==2.6.2
django-markdown==0.8.4
django-haystack==2.4.0
elasticsearch==1.7.0
